requirements.txt
langchain
langchain-openai
langchain-core
langchain-community
faiss-cpu
chromadb
pypdf
python-dotenv
streamlit
tavily-python

#langchain langchain-openai langchain-community chromadb pypdf python-dotenv streamlit faiss-cpu